refactor: unify WORKSPACE_FILES into shared src/config/workspacePaths.ts
Move WORKSPACE_FILES definition to src/config/workspacePaths.ts
(browser-compatible, no node:path dependency) and re-export from
server/workspace/paths.ts. Eliminates the duplicated path strings
between server and frontend.
